requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.23.3
openapi-codec==1.3.2
packaging==21.3
pandas==1.5.0
Pillow==9.1.1
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
zipp==3.8.0
